Revert "FEA: Realiza a criação de formulário com requisição Ajax, ser reavaliado o upload de imagens(BUG)"
