+++ composer.json
@@ -1,7 +1,6 @@
 {
-    "name": "ander-robson/dev-admin",
+    "name": "ander/dev-admin",
     "description": "Painel adiministrativo",
-    "minimum-stability": "stable",
     "type": "project",
     "autoload": {
         "psr-4": {
@@ -33,5 +32,6 @@
             "name": "Ander Robson",
             "email": "ander.robson@hotmail.com"
         }
-    ]
+    ],
+    "minimum-stability": "stable"
 }
