+++ engine/routes.php
@@ -15,9 +15,7 @@
     $router->get("{slug_post}", "Web:slugPost");
     $router->get("pages/{page}", "Web:pages");
     $router->get("pages/{page}/{function}", "Web:pages");
-    $router->get("pages/{page}/{function}/{slug}", "Web:pages");
     $router->post("pages/{page}/{function}", "Web:pages");
-    $router->post("pages/{page}/{function}/{slug}", "Web:pages");
     $router->post("delete", "Web:delete", 'form.delete');
     $router->get("banner", "Web:banner");
