+++ source/Controllers/Controller.php
@@ -7,8 +7,8 @@ namespace Source\Controllers;
 use CoffeeCode\Optimizer\Optimizer;
 use CoffeeCode\Router\Router;
 use League\Plates\Engine;
+use Theme\Pages\Home\HomeModel;
 use Source\Models\Configures;
-use Theme\Pages\User\UserModel;
 
 /**
  * Class Controller
@@ -25,7 +25,7 @@ abstract class Controller
     /** @var Optimizer */
     protected $seo;
 
-    /** @var UserModel */
+    /** @var HomeModel */
     protected $user;
 
     /** @var Configures */
@@ -37,21 +37,19 @@ abstract class Controller
      */
     public function __construct($router)
     {
+        $this->configures = new Configures();
         $this->router = $router;
         $this->view = Engine::create(ROOT . DS . "theme" . DS . "pages", "php");
         $this->view->addData(["router" => $this->router]);
 
         $this->seo = new Optimizer();
+        $this->seo->openGraph(SITE['NAME'], SITE['LOCALE'], "article")
+            ->publisher(SOCIAL["FACEBOOK_PAGE"], SOCIAL["FACEBOOK_AUTHOR"])
+            ->twitterCard(SOCIAL["TWITTER_CREATOR"], SOCIAL["TWITTER_SITE"], SITE["DOMAIN"])
+            ->facebook(SOCIAL["FACEBOOK_APP_ID"]);
 
-        $facebookInformation = $this->getConfigure("facebook_login");
-        if (! empty($facebookInformation)) {
-            $this->seo->openGraph(SITE['NAME'], SITE['LOCALE'], "article")
-                ->publisher(SOCIAL["FACEBOOK_PAGE"], SOCIAL["FACEBOOK_AUTHOR"])
-                ->twitterCard(SOCIAL["TWITTER_CREATOR"], SOCIAL["TWITTER_SITE"], SITE["DOMAIN"])
-                ->facebook($facebookInformation->clientId);
-        }
-
-        if (! empty($_SESSION["user"]) && $this->user = (new UserModel())->findById($_SESSION['user'])) {
+        if (! empty($_SESSION["user"])) {
+            $this->user = (new HomeModel())->findById($_SESSION["user"]);
             $this->view->addData(['user' => $this->user]);
         }
     }
