+++ theme/pages/home/model.php
@@ -1,8 +1,13 @@
 <?php
 
-namespace Theme\Pages\Home;
+    namespace Theme\Pages\Home;
 
-class HomeModel
-{
+    use CoffeeCode\DataLayer\DataLayer;
 
-}
+    class HomeModel extends DataLayer
+    {
+        public function __construct()
+        {
+            parent::__construct("users", ["first_name", "last_name"]);
+        }
+    }
